- feat: effects, key 파일 분리
