+++ app/recoils/sample/user/selector.ts
@@ -1,8 +1,9 @@
-import { selector, selectorFamily } from 'recoil';
+import { selector } from 'recoil';
 import { userState } from './atom';
+import { RECOIL_KEYS } from '../key.d';
 
 export const userTagText = selector({
-  key: 'userTagText',
+  key: RECOIL_KEYS.userTagText,
   get: ({ get }) => {
     const { tagList } = get(userState);
     return '#' + tagList.join(' #');
